Extracts the element pasting code to another class, the ElementPaster
Adds a new interface, Paster
